Numerous changes
-Added a functioning menu bar!
This took an embarrassingly long time but it works! The only thing we
need to adjust is the log out / cookie stuff because the menu bar uses
links not buttons so when the user is logged out by the cookie it still
is letting you select stuff.  I think we can fix this just by putting an
if !isset cookie on each page but we'll see.

-Moved things
I moved over all your map stuff to a new page and to the best of my
knowledge it all still works.

-Got all CSS in external sheet
This one speaks for itself but YAY